Move assert.h into 'internal' folder
