+++ src/adiar/bdd/apply.cpp
@@ -4,11 +4,10 @@
 #include <adiar/file_stream.h>
 #include <adiar/file_writer.h>
 
+#include <adiar/internal/assert.h>
 #include <adiar/internal/levelized_priority_queue.h>
 #include <adiar/internal/product_construction.h>
 #include <adiar/internal/tuple.h>
-
-#include <adiar/assert.h>
 
 namespace adiar
 {
