+++ src/adiar/bdd/if_then_else.cpp
@@ -3,11 +3,10 @@
 #include <adiar/file_stream.h>
 #include <adiar/file_writer.h>
 
+#include <adiar/internal/assert.h>
 #include <adiar/internal/levelized_priority_queue.h>
 #include <adiar/internal/tuple.h>
 #include <adiar/internal/util.h>
-
-#include <adiar/assert.h>
 
 namespace adiar
 {
