+++ src/adiar/bdd/negate.cpp
@@ -3,8 +3,6 @@
 #include <adiar/file_stream.h>
 #include <adiar/file_writer.h>
 
-#include <adiar/assert.h>
-
 namespace adiar
 {
   bdd bdd_not(const bdd &in)
