+++ src/adiar/bdd/quantify.cpp
@@ -6,10 +6,9 @@
 #include <adiar/file_stream.h>
 #include <adiar/file_writer.h>
 
+#include <adiar/internal/assert.h>
 #include <adiar/internal/quantify.h>
 #include <adiar/internal/tuple.h>
-
-#include <adiar/assert.h>
 
 namespace adiar
 {
