+++ src/adiar/internal/intercut.h
@@ -7,9 +7,8 @@
 #include <adiar/file_stream.h>
 #include <adiar/file_writer.h>
 
+#include <adiar/internal/assert.h>
 #include <adiar/internal/levelized_priority_queue.h>
-
-#include <adiar/assert.h>
 
 namespace adiar
 {
