+++ src/adiar/internal/reduce.h
@@ -5,10 +5,9 @@
 #include <adiar/file_stream.h>
 #include <adiar/file_writer.h>
 
+#include <adiar/internal/assert.h>
 #include <adiar/internal/levelized_priority_queue.h>
 #include <adiar/internal/sorter.h>
-
-#include <adiar/assert.h>
 
 #include <adiar/statistics.h>
