+++ src/adiar/zdd/build.cpp
@@ -3,9 +3,8 @@
 #include <adiar/file_stream.h>
 #include <adiar/file_writer.h>
 
+#include <adiar/internal/assert.h>
 #include <adiar/internal/build.h>
-
-#include <adiar/assert.h>
 
 namespace adiar
 {
