+++ src/adiar/zdd/count.cpp
@@ -1,9 +1,9 @@
 #include <adiar/zdd.h>
 
 #include <adiar/data.h>
-#include <adiar/internal/count.h>
 
-#include <adiar/assert.h>
+#include <adiar/internal/assert.h>
+#include <adiar/internal/count.h>
 
 namespace adiar
 {
